backend/requirements.txt
# pip3 install -r requirements.txt

fastapi
uvicorn
requests
python-dotenv
